backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
safetensors>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
tensorboard>=2.15.0
optuna>=3.4.0
Pillow>=10.0.0
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.0.0
transformers
accelerate
qwen-vl-utils
